Add support for all 4 light channels(sun, r, g, b). Also increase resolution of each channel from previously 32 to now 256.

+++ cubyz-common/src/io/cubyz/blocks/BlockInstance.java
@@ -14,7 +14,7 @@ public class BlockInstance {
 	private Vector3i pos;
 	private World world;
 	public boolean neighborUp, neighborDown, neighborEast, neighborWest, neighborNorth, neighborSouth;
-	public byte light;
+	public int light;
 	
 	public World getWorld() {
 		return world;
